refactor OpenID4VPContext
